Add files via upload
The Pickle Files are taken from a bitbucket Repository - https://bitbucket.org/jadslim/german-traffic-signs/src/master/. 
These are then unpickled as written in the code. The traffic signs are divided into training, validation and Test data. 

The accuracy achieved with the model can be seen by opening the images in the repository.
Fit Generation Method was also implemented to increase the overall accuracy. Feel free to check the model on random traffic signs.

Summary:
1. Training Accuracy - 97%
2. Validation Accuracy - 98.87%
3. Test Accuracy - 97%

The accuracy can be further improved by using either more number of Hidden Conv2D layers and/or by increasing the number of filters or epochs.